requirements.txt
sentence-transformers==3.0.1
qdrant-client==1.9.1
groq>=0.13.0
httpx<0.28.0
python-dotenv==1.0.1
requests==2.32.3